+++ q1simulator/q1sequencer.py
@@ -394,8 +394,13 @@ class Q1Sequencer(InstrumentChannel):
     def set_trigger_events(self, events):
         self._trigger_events = events
 
-    def plot(self):
-        self.rt_renderer.plot(self._v_max, plot_label=self.label)
+    def plot(self, t_min=None, t_max=None):
+        self.rt_renderer.plot(
+            self._v_max,
+            plot_label=self.label,
+            t_min=t_min,
+            t_max=t_max,
+            )
 
     def print_registers(self, reg_nrs=None):
         self.q1core.print_registers(reg_nrs)